Upgrade dependencies, refactor to es6
This patch upgrades the site to React 0.14, along with minor upgrades
to most all other dependencies. A sourcemap is now generated for the
javascript, eslint is now used instead of jshint/jscs, es6 syntax is
used throughout, and the code has been simplified. Flux was removed
because this use-case was too simple for Flux to be useful.